backend/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
six==1.16.0
Werkzeug==2.2.2
zipp==3.11.0
